commenting unwanted code

+++ elastalert/ruletypes.py
@@ -1041,8 +1041,8 @@ class ErrorRateRule(BaseAggregationRule):
     def calculate_err_rate(self,payload):
         for timestamp, payload_data in payload.iteritems():
             if int(payload_data['total_count']) > 0:
-                rate = int(payload_data['error_count'])/int(payload_data['total_count'])
-                rate = (rate*100)/self.rules['sampling']
+                rate = float(payload_data['error_count'])/float(payload_data['total_count'])
+                rate = float(rate*100)/float(self.rules['sampling'])
                 rate = rate*100
                 if 'threshold' in self.rules and rate > self.rules['threshold']:
                     match = {self.rules['timestamp_field']: timestamp, 'error_rate': rate}
